Auto-start Media Monkey and better error reporting!
Added new parameter to config file and code to allow the program to
start Media Monkey automatically when completed switching playlists.
This was because I would always wait for the program to complete to open
MM, so why not automate it?

Also, this program would crash if MM were open while it tried to access
the database. That error is reported better now.